Resolves conflicts
